requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-standardized-errors==0.13.0
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.1
